pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sharepoint-checker"
version = "0.1.0"
description = "SharePoint Online tenant-wide folder/file structure validator"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27",
    "msal>=1.29",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "tenacity>=8.3",
    "typer>=0.12",
    "anyio>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",
    "coverage[toml]>=7.4",
]

[project.scripts]
sp-checker = "sharepoint_checker.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/sharepoint_checker"]
omit = ["tests/*"]
